fix cuda

+++ onnxruntime/contrib_ops/cuda/bert/skip_layer_norm.h
@@ -19,7 +19,7 @@ class SkipLayerNorm final : public CudaKernel {
   Status ComputeInternal(OpKernelContext* context) const override;
 
  private:
-  double epsilon_;
+  float epsilon_;
 };
 
 }  // namespace cuda
